backend/requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.2.1
djangorestframework==3.14.0
django-filter==24.1
django-cors-headers==4.3.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.1.0
reportlab==4.0.9
sqlparse==0.5.3
tzdata==2025.2
gunicorn==23.0.0
weasyprint==65.1
djoser==2.2.2
drf-spectacular==0.28.0